+++ src/main/java/amidst/fragment/loader/WorldIconLoader.java
@@ -5,7 +5,7 @@ import amidst.documentation.CalledOnlyBy;
 import amidst.documentation.NotThreadSafe;
 import amidst.fragment.Fragment;
 import amidst.fragment.layer.LayerDeclaration;
-import amidst.mojangapi.world.icon.WorldIconProducer;
+import amidst.mojangapi.world.icon.producer.WorldIconProducer;
 
 @NotThreadSafe
 public class WorldIconLoader extends FragmentLoader {
