+++ src/main/java/amidst/mojangapi/world/WorldBuilder.java
@@ -9,13 +9,13 @@ import amidst.mojangapi.file.nbt.LevelDatNbt;
 import amidst.mojangapi.minecraftinterface.MinecraftInterface;
 import amidst.mojangapi.minecraftinterface.MinecraftInterfaceException;
 import amidst.mojangapi.minecraftinterface.RecognisedVersion;
-import amidst.mojangapi.world.icon.NetherFortressProducer;
-import amidst.mojangapi.world.icon.OceanMonumentProducer;
-import amidst.mojangapi.world.icon.PlayerProducer;
-import amidst.mojangapi.world.icon.SpawnProducer;
-import amidst.mojangapi.world.icon.StrongholdProducer;
-import amidst.mojangapi.world.icon.TempleProducer;
-import amidst.mojangapi.world.icon.VillageProducer;
+import amidst.mojangapi.world.icon.producer.NetherFortressProducer;
+import amidst.mojangapi.world.icon.producer.OceanMonumentProducer;
+import amidst.mojangapi.world.icon.producer.PlayerProducer;
+import amidst.mojangapi.world.icon.producer.SpawnProducer;
+import amidst.mojangapi.world.icon.producer.StrongholdProducer;
+import amidst.mojangapi.world.icon.producer.TempleProducer;
+import amidst.mojangapi.world.icon.producer.VillageProducer;
 import amidst.mojangapi.world.oracle.BiomeDataOracle;
 import amidst.mojangapi.world.oracle.SlimeChunkOracle;
 import amidst.mojangapi.world.player.MovablePlayerList;
