--- src/main/java/amidst/mojangapi/world/icon/WorldIconConsumer.java
@@ -1,5 +1,0 @@
-package amidst.mojangapi.world.icon;
-
-public interface WorldIconConsumer {
-	void consume(WorldIcon worldIcon);
-}
