+++ src/main/java/amidst/mojangapi/world/icon/locationchecker/LocationChecker.java
@@ -1,4 +1,4 @@
-package amidst.mojangapi.world.icon;
+package amidst.mojangapi.world.icon.locationchecker;
 
 public interface LocationChecker {
 	boolean isValidLocation(int x, int y);
