+++ src/main/java/amidst/mojangapi/world/icon/producer/OceanMonumentProducer.java
@@ -1,7 +1,10 @@
-package amidst.mojangapi.world.icon;
+package amidst.mojangapi.world.icon.producer;
 
 import amidst.documentation.NotThreadSafe;
 import amidst.mojangapi.world.coordinates.Resolution;
+import amidst.mojangapi.world.icon.locationchecker.OceanMonumentLocationChecker;
+import amidst.mojangapi.world.icon.type.DefaultWorldIconTypes;
+import amidst.mojangapi.world.icon.type.ImmutableWorldIconTypeProvider;
 import amidst.mojangapi.world.oracle.BiomeDataOracle;
 
 @NotThreadSafe
