+++ src/main/java/amidst/mojangapi/world/icon/producer/PlayerProducer.java
@@ -1,9 +1,10 @@
-package amidst.mojangapi.world.icon;
+package amidst.mojangapi.world.icon.producer;
 
 import java.util.LinkedList;
 import java.util.List;
 
 import amidst.documentation.ThreadSafe;
+import amidst.mojangapi.world.icon.WorldIcon;
 import amidst.mojangapi.world.player.MovablePlayerList;
 import amidst.mojangapi.world.player.Player;
