+++ src/main/java/amidst/mojangapi/world/icon/producer/WorldIconConsumer.java
@@ -1,0 +1,7 @@
+package amidst.mojangapi.world.icon.producer;
+
+import amidst.mojangapi.world.icon.WorldIcon;
+
+public interface WorldIconConsumer {
+	void consume(WorldIcon worldIcon);
+}
