+++ src/main/java/amidst/mojangapi/world/icon/producer/WorldIconProducer.java
@@ -1,4 +1,4 @@
-package amidst.mojangapi.world.icon;
+package amidst.mojangapi.world.icon.producer;
 
 import java.util.List;
 
@@ -6,6 +6,7 @@ import amidst.documentation.AmidstThread;
 import amidst.documentation.CalledOnlyBy;
 import amidst.documentation.NotThreadSafe;
 import amidst.mojangapi.world.coordinates.CoordinatesInWorld;
+import amidst.mojangapi.world.icon.WorldIcon;
 
 @NotThreadSafe
 public abstract class WorldIconProducer {
