+++ src/main/java/amidst/mojangapi/world/icon/type/WorldIconTypeProvider.java
@@ -1,4 +1,4 @@
-package amidst.mojangapi.world.icon;
+package amidst.mojangapi.world.icon.type;
 
 public interface WorldIconTypeProvider {
 	DefaultWorldIconTypes get(int x, int y);
